requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
distro==1.8.0
filelock==3.13.1
FLAML==2.1.1
frozenlist==1.4.0
fsspec==2023.12.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
Jinja2==3.1.2
litellm==1.10.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.2
openai==1.3.7
packaging==23.2
pyautogen==0.2.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.3
zipp==3.17.0
